+++ server/src/services/job-position.service/update-delete-job-position.service.ts
@@ -1,5 +1,6 @@
 import { JobPosition } from "../../models/job-position.model";
-import { NotFoundError } from "../../errors/index";
+import { NotFoundError, AppError } from "../../errors/index";
+import { Employee } from "../../models/employee.model";
 
 export const updateJobPositionService = async (id: string, data: any) => {
   const jobPos = await JobPosition.findByIdAndUpdate(id, data, { new: true, runValidators: true });
@@ -8,6 +9,10 @@ export const updateJobPositionService = async (id: string, data: any) => {
 };
 
 export const deleteJobPositionService = async (id: string) => {
+  const employeeExists = await Employee.exists({ jobPositionId: id });
+  if (employeeExists) {
+    throw new AppError("Cannot delete job position because employees are assigned to it.", 400);
+  }
   const jobPos = await JobPosition.findByIdAndDelete(id);
   if (!jobPos) throw new NotFoundError("Job Position not found");
   return { id };
